app/requirements.txt
Django==3.2
djangorestframework==3.13.1
# Pillow
python-dateutil
psycopg2-binary==2.9.1
gunicorn==20.1.0